Created a custom hook useCheckUser to check if the user is authenticated and if not redirect to login. This makes it easier to reuse on every admin route.
