+++ frontend/src/NewAuthorPage.jsx
@@ -1,6 +1,5 @@
-import { useState, useEffect, useContext } from 'react';
-import { useNavigate } from 'react-router-dom';
-import UserContext from './UserContext';
+import { useState } from 'react';
+import useCheckUser from './useCheckUser';
 
 function NewAuthorPage() {
   const [firstName, setFirstName] = useState(null);
@@ -9,14 +8,8 @@ function NewAuthorPage() {
   const [referido, setReferido] = useState(null);
   const [email, setEmail] = useState(null);
   const [category, setCategory] = useState(null);
-  const { user } = useContext(UserContext);
-  const navigate = useNavigate();
 
-  useEffect(() => {
-    if (user !== undefined && (user === null || user.is_admin === false)) {
-      navigate("/");
-    }
-  }, [user]);
+  useCheckUser();
 
   async function handleSubmit(e) {
     e.preventDefault();
